==> R/fssgam - broad ga bathymetry/fish_fssgam/04_habitat_fish_model_predict.R <==
###
# Project: Parks - Montes omp
# Data:    BRUVS + Habitat data
# Task:    Habitat-Fish modelling + Prediction
# author:  Kingsley Griffin & Claude
# date:    April 2022
##

rm(list=ls())

library(reshape2)
library(mgcv)
library(ggplot2)
library(viridis)
library(raster)
library(dplyr)
library(stringr)
library(sp)
library(sf)

wgscrs  <- CRS("+proj=longlat +datum=WGS84")
sppcrs  <- CRS("+proj=utm +zone=50 +south +datum=WGS84 +units=m +no_defs")     # crs for sp objects

aumpa  <- st_read("data/spatial/shape/AustraliaNetworkMarineParks.shp")           # all aus mpas
nw_mpa <- aumpa[aumpa$ResName %in% c("Montebello"), ]           

# read in
dat1 <- readRDS("data/tidy/dat.maxn.rds")%>%
  glimpse()
dat2 <- readRDS("data/tidy/dat.length.rds")
fabund <- bind_rows(dat1,dat2)                                                  # merged fish data used for fssgam script
prel   <- readRDS("output/spatial_predictions_broad/predicted_relief_raster_ga.rds")           # predicted relief from 'R/habitat/5_krige_relief.R'
proj4string(prel) <- sppcrs                                                     # Kingsley did the relief preds in utm
prel <- projectRaster(prel, crs = wgscrs)                                       # Back to lat long wgs84
plot(prel)

tifs  <- list.files("output/spatial_covariates/", "*.tif", full.names = TRUE)
preds <- stack(tifs)

preddf <- as.data.frame(preds, xy = TRUE, na.rm = TRUE)

# join habitat and relief predictions, fix column names
predsp             <- SpatialPointsDataFrame(coords = cbind(preddf$x, preddf$y), data = preddf)
preddf$mean.relief      <- extract(prel, predsp)
preddf             <- na.omit(preddf)                                           # Removes all rows with NA relief
preddf <- preddf %>%
  dplyr::rename(depth = layer_ga_Z, detrended = layer_ga_detrended, roughness = layer_ga_roughness,
                slope = layer_ga_slope, tpi = layer_ga_tpi)%>%
  glimpse()
fabund <- fabund %>%
  dplyr::rename(depthx = depth, depth = depth_ga)%>%
  glimpse()

# reduce predictor space to fit survey area (needed later)
fishsp <- SpatialPointsDataFrame(coords = cbind(fabund$longitude, 
                                                fabund$latitude), 
                                 data = fabund)
sbuff  <- buffer(fishsp, 10000)

# use formula from top model from FSSGam model selection
#total abundance
m_totabund <- gam(number ~ s(mean.relief, k = 3, bs = "cr"), 
               data = fabund%>%dplyr::filter(response%in%"total.abundance"), 
               method = "REML", family = tw())
summary(m_totabund)

#species richness
m_richness <- gam(number ~ s(mean.relief, k = 3, bs = "cr"), 
                     data = fabund%>%dplyr::filter(response%in%"species.richness"), 
                     method = "REML", family = tw())
summary(m_richness)

#greater than legal size
m_legal <- gam(number ~ s(mean.relief, k = 3, bs = "cr"),  
                  data = fabund%>%dplyr::filter(response%in%"greater than legal size"), 
                  method = "REML", family = tw())
summary(m_legal)

#smaller than legal size
m_sublegal <- gam(number ~ s(depth, k = 3, bs = "cr"),
               data = fabund%>%dplyr::filter(response%in%"smaller than legal size"), 
               method = "REML", family = tw())
summary(m_sublegal)


# predict, rasterise and plot
preddf <- cbind(preddf, 
                "p_totabund" = predict(m_totabund, preddf, type = "response"),
                "p_richness" = predict(m_richness, preddf, type = "response"),
                "p_legal" = predict(m_legal, preddf, type = "response"),
                "p_sublegal" = predict(m_sublegal, preddf, type = "response"))

prasts <- rasterFromXYZ(preddf[, c(1, 2, 9:12)]) 
plot(prasts)

# subset to 10km from sites only
sprast <- mask(prasts, sbuff)                                                   # 10km from sample sites
plot(sprast)
crs(sprast) <- wgscrs
writeRaster(sprast[[3]], "output/fssgam - fish_broad/parks-montes-legal.tif", 
            overwrite = T)
saveRDS(sprast, file = "output/fssgam - fish_broad/montes-fish-spatial_WGS84.rds")

# # to make square
# sprast <- mask(prasts, extent(nw_mpa), inverse = TRUE)                                                   # 10km from sample sites
# plot(sprast)


# tidy and output data
spreddf <- as.data.frame(sprast, xy = TRUE, na.rm = TRUE)

summary(spreddf)

#saveRDS(preddf, "output/broad_fish_predictions.rds")
saveRDS(spreddf, "output/spatial_predictions_broad/site_fish_predictions.rds")
